Support using 'gf' to go to an included view
This enables you to use the keystroke 'gf' to go to the included or
extended view under the cursor, i.e. @include('filename'),
@extends('filename').

The semicolon after the path makes it possible to go to the file, even
if your PWD is somewhere deeper in the project, e.g.
resources/views/frontend/pages (see :help file-searching).