WIP: add indent for multiline include statements
I only added this for `@include`, not for other directives, because I
think it only makes sense for `@include` to be on multiple lines (i.e.
when you pass it a large array)

It doesn't work properly yet when there's a function call in the include
statement.  This is because it falsely recognizes the closing
parenthesis of the function call as the closing parenthesis of the
include statement.