Remove invalid conditional
This conditional contained a syntax error (one `=` instead of two) and
therefore got ignored by vim.  When I try to 'fix' the syntax error, the
indentation gets all messed up.  So I'll just remove it as it isn't
needed.

I also removed the `cindent` variable that only was used in that
conditional.