Add failing test for @if directive inside html tag
I realise you probably won't want to merge this without the fix. If I had the first clue how to write the fix I would have submitted that too, but thought this would be a good way to alert you to the problem anyhow. In the meantime, I will start reading up on how to do this to submit the fix.